refactor(*): delete files and sort hooks files
delete helper
