--- src/hooks/globalEvents.js
@@ -1,16 +1,0 @@
-import { Router, LinkEvent } from './router.js'
-
-export function GlobalEventsHook() {
-  handleHashChange()
-  handleLinkEvent()
-}
-
-function handleLinkEvent() {
-  document.addEventListener('click', LinkEvent)
-}
-
-function handleHashChange() {
-  const preventDefault = () => {}
-  window.addEventListener('hashchange', () => Router({preventDefault}))
-  Router({ preventDefault })
-}
